requirements.txt
requests
cryptography
python-nmap
ipadress